internal: Move VCS into its own crate. (#167)
* Move files.

* Fix deps.
