+++ crates/vcs/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "moon_vcs"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+moon_config = { path = "../config" }
+moon_error = { path = "../error" }
+moon_utils = { path = "../utils" }
+async-trait = "0.1.53"
+ignore = "0.4.18"
+regex = "1.5.6"
+thiserror = "1.0.31"
+tokio = "1.18.2"
