+++ crates/vcs/src/lib.rs
@@ -1,0 +1,9 @@
+mod errors;
+mod git;
+mod loader;
+mod svn;
+mod vcs;
+
+pub use errors::VcsError;
+pub use loader::*;
+pub use vcs::*;
