+++ crates/workspace/src/lib.rs
@@ -3,12 +3,10 @@ mod action_runner;
 mod actions;
 mod dep_graph;
 mod errors;
-mod vcs;
 mod workspace;
 
 pub use action::{Action, ActionStatus};
 pub use action_runner::ActionRunner;
 pub use dep_graph::DepGraph;
 pub use errors::WorkspaceError;
-pub use vcs::TouchedFiles;
 pub use workspace::Workspace;
